test with normal function instead of subprocess
